Make verify a standalone function.
Also moves verify and get_audience to base.py and makes them
importable from django_browserid directly. Their old references
have been deprecated and will be removed after a reasonable(?)
amount of time.
